# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'Tijori API'
  description: 'Tijori family networking platform API — phone auth, family graph, encrypted chat & media. Modules are grouped by tag in Swagger UI.'
  version: 1.0.0
servers:
  - url: /api/v1
    description: 'API v1'
paths:
...
  /storage/quota:
    get:
      tags:
        - StoragePlans
      summary: 'Get my storage quota and usage'
      operationId: storageQuota
      responses:
        200:
          description: 'Quota summary'
      security:
        -
          bearerAuth: []
  /storage/plans:
    get:
      tags:
        - StoragePlans
      summary: 'List active storage plans'
      operationId: storagePlansCatalog
      responses:
        200:
          description: 'Plan catalog'
      security:
        -
          bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Use token from POST /auth/login or POST /auth/register'
      bearerFormat: Sanctum
      scheme: bearer
...